fix(testables): Prevent using nextest for doctests
When `cargo-nextest` is detected, it becomes the executable to trigger
`testable` runnables.

Unfortunally, `nextest` does not support doc tests. This causes the
doctest runnables to fail when the binary is available on the path.

This commit prevents the use of 'nextest' for 'doctest' runnables,
falling back to the `cargo test --doc` executable in this case.

This way, we can do `:RustLsp testables` and run both `#[tests]` and
`doctests` properly.