feat(neotest): Expose doctests on Neotest summary window
The current neotest integration reports any 'test' runnable as a valid
position, except for 'doctests'.

This commit introduces logic necessary to also expose 'doctests' as part
of the summary, so it can be visualized and trigger.

It tries to approach it with the minimal changes, so exported labels
follows the same name convention as test runnables, without indication
they are doctests on the summary window. This can be changed if
necessary.

In order for neotest to consider these items part of the tree, we must
also return a 'pos' with `pos.type = 'file'`, otherwhise those test
items would not have a parent, and it prevents being displayed. This is
done by reporting a 'file' always, but I'm not sure what is the reason
behind returning only if `#namespaces > 0`.