fix(lsp): renderDiagnostics and explainError skip valid diagnostics
Multiple diagnostics can exist at the same cursor location but
vim.diagnostic.get_next() will only return one. If this diagnostic is
not the correct type for what we are doing (rendering
diagnostics/explaining errors), we would start searching for another
diagnostic further in the file. This could skip valid diagnostics if
they weren't the one returned by get_next().

We now search through the filtered list of diagnostics that meet our
criteria to see if there is a valid one at that location that we can use
instead, and only continue onto further cursor locations if none of them
do.