feat(config): Allow overriding the root_dir
I'm working in a repository with git submodules, which leads to nested
cargo workspaces. When working in the top-level repository, I always
want rust-analyzer to use the top-level repository as the root, so that
I don't end up with multiple overlapping instances of rust-analyzer.

This is a niche use-case, but it would be nice to have a way to override
the default behavior.

This is loosely modeled after nvim-lspconfig's `root_dir` option:
https://github.com/neovim/nvim-lspconfig/blob/94513a5b246cf32a8f87ca714af50911df63351c/CONTRIBUTING.md#adding-a-server-to-lspconfig

- **Why `get_root_dir` and not `root_dir`?** `RustaceanLspClientConfig`
  is a subtype of `vim.lsp.ClientConfig`. This is taken advantage of
  when constructing the config for each individual client. The
  `vim.lsp.ClientConfig` type has a conflicting `root_dir` field that's
  typed as a string (not a function).

- **Why does it not take `bufnr`, like nvim-lspconfig's API?** There's
  no way to get the `bufnr` value in the neotest adapter. It's also
  probably not needed in most cases.